% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicating.R
\name{indicating}
\alias{indicating}
\title{Checks whether true value is between CI limits.}
\usage{
indicating(UPPER, LOWER, trueVal)
}
\description{
This function takes a true value and checks whether the value is between
the upper and lower limits of a given Confidence Interval.
}
