% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueMCvalues.R
\name{trueMCvalues}
\alias{trueMCvalues}
\title{True parameter values.}
\usage{
trueMCvalues(ID = c("sim_act"), keyword)
}
\arguments{
\item{ID}{An identifier to select the appropriate simulation scenario.}

\item{keyword}{The parameter that needs to be selected.}
}
\description{
This function returns true parameter value for the Monte-Carlo simulations
It takes an ID for the scenario (e.g. data with activation) and keyword as argument.
It is used to define global variables once (here). The scripts running the
simulations, analyses, visualizations etc., rely on this function.
The idea is to reduce the amount of copying global variables between scripts.
}
\details{
For ID, we have the following possibilities at the moment:
\itemize{
\item act_sim
}
}
